+++ src/include/helpers.h
@@ -19,7 +19,7 @@ int get_monitor_count(Display*, const Window);
 void add_args_to_command(char**, const int, char**);
 void add_radius_to_args(char*, const int, const int);
 void die(const char*, const char*, const int, const uint8_t);
-void get_monitor_centers(Display*, const Window, const int, int*, int*);
+void get_monitor_centers(Display*, const Window, const size_t, int*, int*);
 void warn(const char*, const char*, const int);
 void warn_if_errno(const char*, const char*, const int);
